+++ CHANGES.md
@@ -1,2 +1,6 @@
+### 1.1.0 ###
+Added: NodeMCU-Tool to use the connector as well as all CLI functions programmatic
+Changed: the bin tool has been separated in two files
+
 ### 1.0.0 ###
 Initial public release